moved pbc into pairwise
